# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/astrologer/unblock-user/{userId}": {
      "delete": {
        "operationId": "AstrologerBlockingController_unblockUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerBlocking"
        ]
      }
    },
...
    "/api/v1/admin/admins/{adminId}": {
      "get": {
        "operationId": "AdminManagementController_getAdminDetails",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      },
      "patch": {
        "operationId": "AdminManagementController_updateAdmin",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      },
      "delete": {
        "operationId": "AdminManagementController_deleteAdmin",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      }
    },
...
    "/api/v1/admin/payments/recharge-packs/{amount}": {
      "delete": {
        "operationId": "AdminRechargePacksController_deletePack",
        "parameters": [
          {
            "name": "amount",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminRechargePacks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateAdminDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}